+++ src/services/Environment.js
@@ -3,6 +3,8 @@ export const EnvironmentDev = {
     production: false,
     apiUrl: 'http://127.0.0.1:8000/api',
     apiUrlNodeMicroservice: 'http://127.0.0.1:9000',
+    apiUrlBillingMicroservice: 'http://127.0.0.1:9001',
+    STRIPE_PUBLIC_KEY: 'pk_test_AxhlK7IudWRlQCa4azz4WElP00QVVE6SNM',
 };
 
 export const EnvironmentStaging = {
@@ -10,6 +12,8 @@ export const EnvironmentStaging = {
     production: false,
     apiUrl: 'https://atila-7-staging.herokuapp.com/api',
     apiUrlNodeMicroservice: 'https://tgrr8bis30.execute-api.us-east-1.amazonaws.com/staging',
+    apiUrlBillingMicroservice: 'https://noitlizkka.execute-api.us-east-1.amazonaws.com/staging',
+    STRIPE_PUBLIC_KEY: 'pk_test_AxhlK7IudWRlQCa4azz4WElP00QVVE6SNM',
 };
 
 export const EnvironmentProd = {
@@ -17,6 +21,8 @@ export const EnvironmentProd = {
     production: true,
     apiUrl: 'https://atila-7.herokuapp.com/api',
     apiUrlNodeMicroservice: 'https://yhpl8yynpk.execute-api.us-east-1.amazonaws.com/prod',
+    apiUrlBillingMicroservice: 'https://eiqnyq1jx6.execute-api.us-east-1.amazonaws.com/prod',
+    STRIPE_PUBLIC_KEY: 'pk_live_AgHuMdhPDSk5eKrlwoHCnZnR00uIrfIHkM',
 };
 
 let Environment = {};